se añadio a la seccion acerca de el logo de telegram y a los restantes iconos se les actualizo la url a la real , la barra de navegacion agrego eluso de localstorage para poder cargar correctamente la posicioncuando se actualice la pagina y su uso en barra lateral , en common se añadio sececcionar y opcion para dicha  tarea, se creo paginacion para el final de la pagina pero hay que mejorarlo , se creo la interface de bebidas, se añadio un parametro a la interface de pedido cantRestante,  se elemino un parametro cantidad de la interface PedidoCrear en relacion con la eliminacion de esacolumna en la base de datos, notificaciones implemento el select , se cambio un parametro de pedido por otro nombre, se cambiaron las rutas y los nombres de los componentes sopa y parrilla por bebidas y postres y se arreglo definitivamente el  contexto para las funciones handlePedido yHandleSubPedido

+++ src/VistaPrincipal/PaginasVistaPrincipal/PaginaSecundaria/UI-Secundaria/Pedido.tsx
@@ -1,14 +1,15 @@
 import { RiCheckboxCircleFill, RiCheckboxCircleLine, RiDeleteBin6Fill, RiDeleteBin6Line } from "react-icons/ri";
 import { useContext, useState } from "react";
 import { Contexto } from "../../../../Contexto";
-import { deletePedido } from "../Secundaria-Servicios";
+import { deletePedido, getUnPedido } from "../Secundaria-Servicios";
 import { Cargando } from "../../../../common";
 import { updatePlatoCaliente, updatePlatoCalienteByPedido } from "../../UI";
+import { updateBebidaByPedido } from "../..";
 
 
 interface Props extends React.HTMLAttributes<HTMLDivElement> { 
     identificador:string;
-    identificadorPlato:string;
+    identificadorProducto:string;
     descripcion:string;
     cantidad:number;
     montoTotal:number;
@@ -16,7 +17,7 @@ interface Props extends React.HTMLAttributes<HTMLDivElement> {
     tipoProducto:string;
  }
 
-export function Pedido({identificador,identificadorPlato,descripcion,cantidad,montoTotal,imagen,tipoProducto}:Props){
+export function Pedido({identificador,identificadorProducto,descripcion,cantidad,montoTotal,imagen,tipoProducto}:Props){
   
     
 
@@ -26,12 +27,14 @@ export function Pedido({identificador,identificadorPlato,descripcion,cantidad,mo
         let color = "";
         if(tipoProducto == "Plato Caliente"){
               color = "text-[#ec7c6a]";
+        } else if(tipoProducto == "Bebida"){
+              color = "text-emerald-500";
         }
         return color;
       });
       
 
-    const { escuchaPedidos,setEscuchaPedidos,HanddleDevolverTodosLosPlatos,setEscuchaPlatoCaliente } = useContext(Contexto);
+    const { escuchaPedidos,setEscuchaPedidos,setEscuchaBebidas,setEscuchaPlatosCalientes } = useContext(Contexto);
 
       if(escuchaPedidos){
 
@@ -101,28 +104,42 @@ export function Pedido({identificador,identificadorPlato,descripcion,cantidad,mo
         </div>
         <div className="grid grid-cols-6">
             <p className={`bg-[#1F1D2B] col-span-5 py-1 px-4 rounded-lg outline-none ${colorProducto}`}>{tipoProducto}</p>
-            {/* <button
-                onMouseEnter={() => setguardarNota(true)}
-                onMouseLeave={() => setguardarNota(false)}
-                className={`mx-1 rounded-lg border text-blue-500 border-blue-500 ${guardarNota? 'bg-blue-500 text-black border-none': ''} flex items-center justify-center`}
-                >
-                {guardarNota ? <RiCheckboxCircleFill/> : <RiCheckboxCircleLine/>}
-            </button> */}
             <button
                 onMouseEnter={() => seteliminarNota(true)}
                 onMouseLeave={() => seteliminarNota(false)}
                 disabled={escuchaPedidos}
                 className={`mx-1 rounded-lg border text-red-500 border-red-500 ${eliminarNota? 'bg-red-500 text-yellow-300 border-none': ''} flex items-center justify-center`}
-                onClick={ async ()=> {
-                   await deletePedido(identificador);
-                   await updatePlatoCalienteByPedido(identificadorPlato,cantidad).then(()=> {
+                onClick={ 
+                    async ()=> {
+                        await getUnPedido(identificador)
+                        .then( (data)=> {
+                            const tipoProducto = data.tipoProducto;
 
-                       HanddleDevolverTodosLosPlatos(identificadorPlato,cantidad);
+                            if(tipoProducto == "Plato Caliente"){
+                          updatePlatoCalienteByPedido(identificadorProducto,cantidad);
+                          deletePedido(identificador)
+                         .then( ()=> {
+
+                    //    HanddleDevolverTodosLosPlatos(identificadorPlato,cantidad);
                        setEscuchaPedidos(true);
-                       setEscuchaPlatoCaliente(true);
-                   });
-                    // setEscuchaPlatosCalientes(true);
-                } }
+                       setEscuchaPlatosCalientes(true);
+                   } )
+                            } else if(tipoProducto ==  "Bebida"){
+                                  console.log("bebidas");
+                                  updateBebidaByPedido(identificadorProducto,cantidad);
+                                  deletePedido(identificador)
+                         .then( ()=> {
+
+                    //    HanddleDevolverTodosLosPlatos(identificadorPlato,cantidad);
+                       setEscuchaPedidos(true);
+                       setEscuchaBebidas(true);
+                   } )
+                                  
+                            }
+                            
+                        } )
+                } 
+            }
                 >
                 {eliminarNota ? <RiDeleteBin6Fill/> : <RiDeleteBin6Line/>}
             </button>
